requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
jinja2==3.1.3
openai==1.61.1
firebase-admin==6.4.0
python-dotenv==1.0.1
httpx==0.27.0
loguru==0.7.2
pydantic==2.4.2
python-multipart==0.0.20
textblob==0.19.0
beautifulsoup4==4.13.4
spacy==3.8.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl